Upload file in API

+++ src/Hackaton.Api/Data/Entities/File.cs
@@ -5,7 +5,7 @@ namespace Hackaton.Api.Data.Entities;
 public class File : Entity
 {
     public string Name { get; set; }
-    public int SizeInBytes { get; set; }
+    public long SizeInBytes { get; set; }
     public string? Url { get; set; }
     public string ContentType { get; set; }
     public ProcessStatusEnum ProcessStatusId { get; set; }
@@ -15,7 +15,7 @@ public class File : Entity
     {
     }
 
-    public File(string name, int sizeInBytes, string? url, string contentType, ProcessStatusEnum processStatusId)
+    public File(string name, long sizeInBytes, string? url, string contentType, ProcessStatusEnum processStatusId)
     {
         Name = name;
         SizeInBytes = sizeInBytes;
